dom/tests/mochitest/webapps/test_bug_795164.xul

<?xml version="1.0"?>

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->

<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Mozilla Bug 795164">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="head.js"/>
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=741549"
     target="_blank">Mozilla Bug 795164</a>
  </body>
<script>

var url1 = "http://test1.example.com/chrome/dom/tests/mochitest/webapps/apps/basic.webapp";
var url2 = "http://test2.example.com/chrome/dom/tests/mochitest/webapps/apps/basic.webapp";

var app1, app2;

// Keeping track of the number of times `mozApps.mgmt.onuninstall` gets triggered
var uninstallCount = 0;

var steps = [
  listenToUninstall,
  installTwoApps,
  uninstallTwoApps,
  checkUninstallCount,
  cleanUp
];

runAll(steps);

function listenToUninstall(next) {
  navigator.mozApps.mgmt.onuninstall = function onUnInstall() {
    uninstallCount++;
  };
  next();
}

function installTwoApps(next) {
  confirmNextInstall();
  navigator.mozApps.install(url1, null).onsuccess = function onInstall(evt) {
    app1 = evt.target.result;
    confirmNextInstall();
    navigator.mozApps.install(url2, null).onsuccess = function onInstall(evt) {
      app2 = evt.target.result;
      next();
    };
  };
}

function uninstallTwoApps(next) {
  navigator.mozApps.mgmt.uninstall(app1).onsuccess = function onUninstallApp1() {
    navigator.mozApps.mgmt.uninstall(app2).onsuccess = function onUninstallApp2() {
      next();
    };
  };
}

function checkUninstallCount(next) {
  is(uninstallCount, 2, "mgmt.onuninstall got triggered only twice");
  next();
}

function cleanUp(next) {
  navigator.mozApps.mgmt.onuninstall = null;
  next();
}

</script>
</window>
